refactor(ui): relocate SyncStatusCard from Home to top of Sync
Home tab's at-a-glance stats card was taking meaningful vertical
space directly under the supporter pill, dominating the aesthetic.
Moves the card (and its data plumbing) to the top of the Sync tab
where library-status info naturally belongs.

- SyncStatusInfo + SyncDisplayStatus moved to :feature:sync
- SyncStatusCard composable + helpers moved to :feature:sync/components
- SyncViewModel/SyncUiState gain syncStatus + hasEverSynced, mirroring
  the HomeViewModel flow assembly verbatim (no behavioural changes);
  spotifyConnected / youTubeConnected were already on SyncUiState
- Home tab: SyncStatusCard call site removed; HomeUiState/HomeViewModel
  pruned of syncStatus, spotifyConnected, youTubeConnected, hasEverSynced
  (and the unused totalTracks / totalStorageBytes fields they pulled
  through). MusicData simplified, sourceCountsFlow + syncStatusFlow
  dropped, authStateFlow renamed to promptsFlow now that it only
  carries the Last.fm + Lossless banner prompts. :core:auth dep no
  longer needed on :feature:home.
- Visual placement: above the "Sync" page title on Sync, as the first
  scrollable item — matches the verbatim instruction.

No new functionality; pure relocation. The card renders identically.
